메인 왼쪽 DB연동 및 수정, 트랜잭션 추가, Calendar 기능 업데이트

+++ src/main/java/com/mycompany/webapp/dto/TeamHistory.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class TeamHistory {
 
    
-   int thn;
+   long thn;
    String thtid;
 
    long tha;
